Sped up gameplay via selective floor noise application
Only apply the floor Nosie to a room, when it's being entered. This decreases the lag spike by N-1/N% on average where N is the amount of open pathways.

To improve upon this, go back to updating all the possible room floors and just update less of the cells each frame, splitting the workload between many frames.