Fully optimized Noise Generation
The code now splits the given terrain updates into batches of 100 cells and applies one batch per frame.
As long as it's impossible to enter a new room within (C/100)*N Frames, where C is the number of Cells to be updated in an average room and N is the number of pathways from a room, then this code should make room generation(in its current state) almost undetectable.